Replace writeShadowedClientAndServerModules
